improved following suggestions of es-lint.
